+++ src/edgepi/eeprom/protobuf_assets/proto_files/adc_module.proto
