+++ src/edgepi/eeprom/protobuf_assets/proto_files/calibration_parameters.proto
