+++ src/edgepi/eeprom/protobuf_assets/proto_files/edgepi_module.proto
