+++ src/edgepi/eeprom/protobuf_assets/proto_files/keys.proto
