+++ src/edgepi/eeprom/protobuf_assets/proto_files/rtd_module.proto
