+++ src/edgepi/eeprom/protobuf_assets/proto_files/tc_module.proto
